Initial public release
This board plugs into the Cyclone II mini board I've used in many other projects. It holds a DAC that can produce the X, Y and Z signals necessary to drive a B&W vector monitor along with the associated circuitry to amplify the signals to levels required and provide gain and offset adjustments. There is also space for a ROM to hold the game code and an audio amplifier to drive the speaker. This has been tested with Asteroids Deluxe released by MikeJ of fpgaarcade.